[ioq3] Update ioq3 from upstream

+++ engine/code/client/cl_cin.c
@@ -609,7 +609,7 @@ static unsigned int yuv_to_rgb24( long y, long u, long v )
 	if (g > 255) g = 255;
 	if (b > 255) b = 255;
 	
-	return LittleLong ((r)|(g<<8)|(b<<16)|(255<<24));
+	return LittleLong ((unsigned long)((r)|(g<<8)|(b<<16))|(255UL<<24));
 }
 
 /******************************************************************************
